Remove task count from title in task browser & try to import refresh timer from upstream
The feature of displaying the number of tasks in stock is not needed anymore
as we transfered to header bar and we are not interested in providing this
kind of information there anymore.

Therefore, I removed the parts of code which took care of displaying this
information.

Fixes #117 and fixes #158